🐛 FIX:

+++ backend/database/migrations/tenant/2024_05_14_180138_add_branch_id_to_coupons_table.php
@@ -11,11 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('coupons', function (Blueprint $table) {
-            $table->unsignedBigInteger('branch_id')->nullable();
-            $table->foreign('branch_id')->references('id')->on('branches')->onDelete('cascade');
-
-        });
+        if (!Schema::hasColumn('coupons', 'branch_id')) {
+            Schema::table('coupons', function (Blueprint $table) {
+                $table->unsignedBigInteger('branch_id')->nullable();
+                $table->foreign('branch_id')->references('id')->on('branches')->onDelete('cascade');
+            });
+        }
     }
 
     /**
